Add frontend folder; add dots functionality
WOW there are a lot of changes to report. First, I added a frontend
directory, which lets Docker run much faster. I also made some
playability changes to the map and set up the first 2 dots. The rest
just need the specific numbers added in and then the game is ready to be
played! Waiting on @mRaffill for the pics.
Still TODO:
1. Start on invite
2. Audio
3. Phone support
4. More polishing, easter eggs, etc